Set initial focus of splash screen to "Okay" button
On the splash screen, changed the initial focus from the "exit" button to the "okay" button.

This causes an Enter key press to continue with manual installation, unless the focus is changed
by pressing arrow keys.  I decided to keep the other buttons "focusable" so that they can be
navigated by the arrow keys.

My previous method of setting a default for enter
self.set_default(button1)
doesn't work when all the UI elements are labels and buttons.  The UI automatically focuses
on one of the buttons, and that button will always be pressed when the user hits enter.